mock api with repository working

+++ src/test/java/com/hashedin/hu/CompOffTest.java
@@ -10,7 +10,7 @@ import static org.junit.Assert.*;
 
 public class CompOffTest {
     LeaveManager manager = new LeaveManager();
-    Employee e1 = new Employee("e1",1,0,Gender.MALE, "2019-01-01");
+    Employee e1 = new Employee("e1", (long) 1,0,Gender.MALE, "2019-01-01");
     @Test
     public void testCompOffNotAddedIfLogWorkLessThan8hrs(){
         e1.addCompoff(LocalDateTime.of(2019, Month.FEBRUARY,1,10,10),LocalDateTime.of(2019,Month.FEBRUARY,1,11,1,1));
